添加正则以及null判断

+++ baseLibs/src/main/java/com/ijustyce/fastandroiddev/baseLib/utils/FileUtils.java
@@ -1,11 +1,11 @@
 package com.ijustyce.fastandroiddev.baseLib.utils;
 
 import android.content.Context;
+import android.content.Intent;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Environment;
-import android.support.annotation.NonNull;
 
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
@@ -24,13 +24,14 @@ import java.io.OutputStream;
 public class FileUtils {
 
     /**
-     *  递归计算文件夹下面的文件的大小，以M为单位,结果保留两位小数(不四舍五入)
+     * 递归计算文件夹下面的文件的大小，以M为单位,结果保留两位小数(不四舍五入)
+     *
      * @param file 文件夹或者文件名
      * @return 文件大小 以M 为单位
      */
-    public static double getDirSize(@NonNull File file) {
+    public static double getDirSize(File file) {
         //判断文件是否存在
-        if (file.exists()) {
+        if (file != null && file.exists()) {
             //如果是目录则递归计算其内容的总大小
             if (file.isDirectory()) {
                 File[] children = file.listFiles();
@@ -39,7 +40,7 @@ public class FileUtils {
                     size += getDirSize(f);
                 return CommonTool.getShortDouble(size / 1024 / 1024);
             } else {//如果是文件则直接返回其大小,以“兆”为单位
-                return CommonTool.getShortDouble(file.length() / 1024/ 1024);
+                return CommonTool.getShortDouble(file.length() / 1024 / 1024);
             }
         } else {
             System.out.println("文件或者文件夹不存在，请检查路径是否正确！");
@@ -49,12 +50,14 @@ public class FileUtils {
 
     /**
      * 获取可用的 文件目录，先尝试sdcard、然后尝试内部存储空间，最后则是data目录了，如果是sdcard，则会创建name这个文件夹
+     *
      * @param context context
-     * @param name 文件夹名字,如果sdcard可用，会在sdcard创建这个目录
+     * @param name    文件夹名字,如果sdcard可用，会在sdcard创建这个目录
      * @return 成功返回true，失败返回false
      */
-    public static String getAvailablePath(@NonNull Context context, @NonNull String name){
+    public static String getAvailablePath(Context context, String name) {
 
+        if (context == null || name == null) return null;
         File f = context.getExternalFilesDir(null);
         if (f == null) {
             f = Environment.getExternalStorageDirectory();
@@ -75,7 +78,9 @@ public class FileUtils {
      * @param newPath 目标文件路径
      * @return 成功返回true，失败返回false
      */
-    public static boolean copyFile(@NonNull String oldPath, @NonNull String newPath) {
+    public static boolean copyFile(String oldPath, String newPath) {
+
+        if (oldPath == null || newPath == null) return false;
         try {
             int bytesum = 0;
             int byteread;
@@ -110,8 +115,9 @@ public class FileUtils {
      * @param fileName assets 下的文件名称
      * @return 成功返回true，失败返回false
      */
-    public static boolean copyDataToSD(@NonNull Context context, @NonNull String toPath, @NonNull String fileName) {
+    public static boolean copyDataToSD(Context context,String toPath, String fileName) {
 
+        if (context == null || toPath == null || fileName == null) return false;
         try {
             InputStream myInput;
             OutputStream myOutput;
@@ -143,7 +149,10 @@ public class FileUtils {
      * @param mBitmap bitmap
      * @param bitName 文件路径 , 必须 .jpg 结尾 比如 /sdcard/tmp/1.jpg
      */
-    public static boolean savBitmapToPng(@NonNull Bitmap mBitmap, @NonNull String bitName) {
+    public static boolean savBitmapToPng(Bitmap mBitmap, String bitName) {
+
+        if (mBitmap == null || bitName == null) return false;
+
         File f = new File(bitName);
         FileOutputStream fOut;
         try {
@@ -174,7 +183,9 @@ public class FileUtils {
      * @param mBitmap bitmap
      * @param bitName 文件路径 , 必须以 .jpg 结尾 比如 /sdcard/tmp/1.jpg
      */
-    public static boolean savBitmapToJpg(@NonNull Bitmap mBitmap, @NonNull String bitName) {
+    public static boolean savBitmapToJpg(Bitmap mBitmap, String bitName) {
+
+        if (mBitmap == null || bitName == null) return false;
 
         File f = new File(bitName);
         FileOutputStream fOut;
@@ -203,24 +214,26 @@ public class FileUtils {
     /**
      * 读取文件
      *
-     * @param file  文本文件
-     * @return  文件内容 发生异常时，返回null
+     * @param file 文本文件
+     * @return 文件内容 发生异常时，返回null
      */
-    public static String readTextFile(@NonNull File file) {
+    public static String readTextFile(File file) {
+
+        if (file == null || !file.exists()) return null;
 
         String text = null;
         InputStream is = null;
         try {
             is = new FileInputStream(file);
             text = readTextInputStream(is);
-        }catch (IOException e){
+        } catch (IOException e) {
             e.printStackTrace();
-        } finally{
+        } finally {
             try {
                 if (is != null) {
                     is.close();
                 }
-            }catch (IOException e){
+            } catch (IOException e) {
                 e.printStackTrace();
             }
         }
@@ -230,11 +243,12 @@ public class FileUtils {
     /**
      * 从流中读取文件
      *
-     * @param is    流对象
-     * @return  文件内容
+     * @param is 流对象
+     * @return 文件内容
      */
-    public static String readTextInputStream(@NonNull InputStream is){
+    public static String readTextInputStream(InputStream is) {
 
+        if (is == null) return null;
         StringBuilder stringBuilder = new StringBuilder();
         String line;
         BufferedReader reader = null;
@@ -243,13 +257,13 @@ public class FileUtils {
             while ((line = reader.readLine()) != null) {
                 stringBuilder.append(line).append("\r\n");
             }
-        } catch (IOException e){
+        } catch (IOException e) {
             e.printStackTrace();
-        }finally {
+        } finally {
             if (reader != null) {
                 try {
                     reader.close();
-                }catch (IOException e){
+                } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
@@ -260,27 +274,26 @@ public class FileUtils {
     /**
      * 将文本内容写入文件
      *
-     * @param file  要写入的文件
-     * @param str   要写入的内容
+     * @param file 要写入的文件
+     * @param str  要写入的内容
      */
-    public static void writeTextFile(@NonNull File file, @NonNull String str) {
+    public static void writeTextFile(File file, String str) {
 
+        if (file == null || !file.exists() || str == null) return;
         DataOutputStream out;
         try {
             out = new DataOutputStream(new FileOutputStream(file));
             out.write(str.getBytes());
             out.close();
-        } catch (IOException e){
+        } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
     /**
      * 遍历删除文件 如果是目录，则删除目录下的一切内容，目录不会删除，如果是文件，则删除文件
-     *
-     * @param path
      */
-    public static void deleteFile(@NonNull String path) {
+    public static void deleteFile(String path) {
 
         ILog.i("===path===", "path is " + path);
         if (StringUtils.isEmpty(path)) {
@@ -305,13 +318,14 @@ public class FileUtils {
 
     /**
      * 返回选择文件时 选定的文件路径
-     * @param context   context
-     * @param uri   onActivityResult里 Intent 对象 getData
-     * @return  如果存在，返回路径，否则返回null
+     *
+     * @param context context
+     * @param uri     onActivityResult里 Intent 对象 getData
+     * @return 如果存在，返回路径，否则返回null
      */
     public static String getPath(Context context, Uri uri) {
 
-        if (context == null || uri == null){
+        if (context == null || uri == null) {
             return null;
         }
         if ("content".equalsIgnoreCase(uri.getScheme())) {
@@ -321,7 +335,7 @@ public class FileUtils {
             try {
                 cursor = context.getContentResolver() == null ?
                         null : context.getContentResolver().query(uri, projection, null, null, null);
-                if (cursor == null){
+                if (cursor == null) {
                     return null;
                 }
                 int column_index = cursor.getColumnIndexOrThrow("_data");
@@ -335,5 +349,29 @@ public class FileUtils {
             return uri.getPath();
         }
         return null;
+    }
+
+    /**
+     * 调用文件管理器显示某个文件
+     * @param path  文件路径
+     * @param context   Context
+     * @return  如果参数有误或者文件不存在，返回0，如果没有文件管理器返回-1，如果成功返回1
+     */
+    public static int showFile(String path, Context context) {
+
+        if (path == null || context == null || !new File(path).exists()){
+
+            return 0;
+        }
+
+        Uri selectedUri = Uri.parse(path);
+        Intent intent = new Intent(Intent.ACTION_VIEW);
+        intent.setDataAndType(selectedUri, "resource/folder");
+
+        if (intent.resolveActivityInfo(context.getPackageManager(), 0) != null) {
+            context.startActivity(intent);
+            return 1;
+        }
+        return -1;
     }
 }
